modularize drizzleRepository
