+++ src/libs/db/schema.ts
@@ -1,8 +1,8 @@
 import { InferSelectModel } from "drizzle-orm";
-import { serial, text, boolean, pgTable } from "drizzle-orm/pg-core";
+import { uuid, text, boolean, pgTable } from "drizzle-orm/pg-core";
 
 export const UserTable = pgTable('users', {
-    id: serial('id').primaryKey(),
+    id: uuid('id').primaryKey().defaultRandom(),
     email: text('email').notNull().unique(),
     password: text('password').notNull(),
     isVerified: boolean('is_verified').notNull().default(false),
