+++ src/libs/repositories/DrizzleRepository.ts
@@ -1,25 +1,75 @@
-import { type PgTable, PgUpdateSetSource } from "drizzle-orm/pg-core";
-import { InferInsertModel, InferSelectModel } from "drizzle-orm";
+import { BaseRepository } from './BaseRepository';
+import { eq, SQL, sql } from 'drizzle-orm';
+import db from '../config/drizzle.config';
+import { PgTable, PgUpdateSetSource } from 'drizzle-orm/pg-core';
+import { Entity, QueryCriteria } from '../interfaces';
 
-export interface QueryCriteria {
-    limit?: number;
-    offset?: number;
-}
+export abstract class DrizzleRepository<M extends PgTable, U extends Entity<string>>
+    implements BaseRepository<U> {
 
-export interface DrizzleRepository<T extends PgTable, ID extends keyof T["$inferSelect"], U> {
-    findAll(): Promise<U[]>; // Return entity type U (e.g., User)
+    protected constructor(
+        protected readonly drizzleDb: typeof db,
+        protected readonly schema: M,
+        protected readonly primaryKey: keyof U
+    ) { }
 
-    findById(id: T["$inferSelect"][ID]): Promise<U | null>; // Return entity type U or null
+    async create(data: Partial<U>): Promise<U> {
+        const transformedData = Object.keys(data).reduce((acc, key) => {
+            acc[key as keyof M["$inferInsert"]] = sql`${data[key as keyof M["$inferInsert"] as keyof U]}`;
+            return acc;
+        }, {} as { [Key in keyof M["$inferInsert"]]: SQL<unknown> });
 
-    update(id: T["$inferSelect"][ID], data: PgUpdateSetSource<T>): Promise<U>; // Return updated entity type U
+        const result = await this.drizzleDb
+            .insert(this.schema)
+            .values(transformedData)
+            .returning();
 
-    create(data: InferInsertModel<T>): Promise<U>; // Return created entity type U
+        return result[0] as unknown as U;
+    }
 
-    delete(id: T["$inferSelect"][ID]): Promise<void>;
+    async findAll(): Promise<U[]> {
+        const result = await this.drizzleDb.select().from(this.schema);
+        return result as U[];
+    }
 
-    list(
-        where?: Partial<U>,  // Use U for the type in the where clause
-        limit?: number,
-        page?: number
-    ): Promise<U[]>; // Return entity type U
+    async findById(id: string): Promise<U | null> {
+        const result = await this.drizzleDb
+            .select()
+            .from(this.schema)
+            .where(eq(sql`${this.primaryKey}`, id));
+        return result.length > 0 ? result[0] as U : null;
+    }
+
+    async update(id: string, data: PgUpdateSetSource<M>): Promise<U> {
+        await this.findById(id);
+        await this.drizzleDb.update(this.schema).set(data).where(eq(sql`${this.primaryKey}`, id));
+        return this.findById(id) as Promise<U>;
+    }
+
+    async delete(id: string): Promise<void> {
+        await this.drizzleDb.delete(this.schema).where(eq(sql`${this.primaryKey}`, id));
+    }
+
+    async list(
+        where?: Partial<U>,
+        criteria?: QueryCriteria
+    ): Promise<U[]> {
+        let query: any = this.drizzleDb.select().from(this.schema);
+
+        if (where) {
+            for (const [key, value] of Object.entries(where)) {
+                query = query.where(eq((this.schema as any)[key], value as any));
+            }
+        }
+
+        if (criteria?.limit) {
+            query = query.limit(criteria.limit);
+        }
+        if (criteria?.offset) {
+            query = query.offset(criteria.offset);
+        }
+
+        const result = await query;
+        return result as U[];
+    }
 }
