+++ src/user-service/utils/generateToken.ts
@@ -1,6 +1,6 @@
 import * as jwt from 'jsonwebtoken';
 
-const generateToken = (id: number): string =>
+const generateToken = (id: string): string =>
     jwt.sign({ user: { id } }, process.env.JWT_SECRET!, {
         expiresIn: '60d',
     });
